api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.5.3
alembic>=1.13.1
pytest>=7.4.4
httpx>=0.26.0  # For testing
python-jose>=3.3.0  # For JWT
passlib>=1.7.4  # For password hashing